requirements.txt
aiosmtpd==1.4.2
atpublic==2.3
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
coverage==5.5
Deprecated==1.2.13
dnspython==2.1.0
expiring-dict==1.1.0
idna==3.3
iniconfig==1.1.1
mariadb==1.0.7
packaging==21.0
pluggy==1.0.0
psutil==5.8.0
py==1.10.0
py-redis==1.1.1
pyparsing==2.4.7
pyspf==2.0.14
pytest==6.2.5
pytest-asyncio==0.15.1
pytest-cov==2.12.1
pytest-order==1.0.0
pytest-services==2.2.1
python-pidfile==3.0.0
redis==4.1.2
requests==2.26.0
sortedcontainers==2.4.0
toml==0.10.2
urllib3==1.26.7
wrapt==1.13.3
zc.lockfile==2.0
